requirements.txt
python-dotenv
slack-bolt
requests
schedule
